+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Master
 
-- Add function variant of `rules` to support execution over multiple packages
+- Add new way of configuring rules via a function to support execution over multiple packages
 
 ## 0.3.2 - Plugins